requirements.txt
asyncio==3.4.3
asyncpg==0.29.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.0.0